requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
coverage==6.4.4
Django==4.1
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.13.1
flake8==5.0.4
importlib-metadata==4.12.0
ipdb==0.13.9
isort==5.10.1
Markdown==3.4.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyHamcrest==2.0.4
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
zipp==3.8.1